Created repo and added files
This is a quantile regressor, an attempt at forecasting quantiles instead of only point forecasts (usually the mean E[Y|X]). I simply turned it into something more suitable for the basic use I had of it and added some documentation of my own to look back at the methods.

The work and design are entirely on https://github.com/ig248/deepquantiles/tree/master/deepquantiles/regressors, for which [blog posts](https://ig248.gitlab.io) accompany the ins and outs of the theory.